pyproject.toml
[project]
name = "widelands-website"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "alabaster==1.0.0",
    "asgiref>=3.11.0",
    "asttokens>=3.0.1",
    "babel>=2.17.0",
    "beautifulsoup4>=4.14.3",
    "bleach>=6.3.0",
    "certifi>=2026.1.4",
    "charset-normalizer>=3.4.4",
    "confusable-homoglyphs==3.3.1",
    "decorator>=5.2.1",
    "django==5.2.10",
    "django-braces>=1.17.0",
    "django-filter>=25.2",
    "django-haystack==3.3.0",
    "django-messages",
    "django-model-utils==5.0.0",
    "django-recaptcha>=4.1.0",
    "django-registration>=5.2.1",
    "django-star-ratings==0.9.2",
    "django-tagging",
    "docutils>=0.22.4",
    "exceptiongroup>=1.3.1",
    "executing>=2.2.1",
    "gunicorn==23.0.0",
    "idna>=3.11",
    "imagesize==1.4.1",
    "ipython>=9.9.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "lxml==6.0.2",
    "markdown>=3.10.1",
    "markupsafe>=3.0.3",
    "matplotlib-inline>=0.2.1",
    "mdx-wikilink-plus==1.4.1",
    "mysqlclient>=2.2.7",
    "packaging>=26.0",
    "parso>=0.8.5",
    "pefile==2024.8.26",
    "pexpect==4.9.0",
    "pillow>=12.0.0",
    "prompt-toolkit>=3.0.52",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydot>=4.0.1",
    "pygments>=2.19.2",
    "pyparsing>=3.3.2",
    "python-magic==0.4.27",
    "python-ptrace==0.9.9",
    "requests>=2.32.5",
    "six==1.17.0",
    "snowballstemmer>=3.0.1",
    "soupsieve>=2.8.2",
    "sphinx>=9.1.0",
    "sphinxcontrib-applehelp==2.0.0",
    "sphinxcontrib-devhelp==2.0.0",
    "sphinxcontrib-htmlhelp==2.1.0",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==2.0.0",
    "sphinxcontrib-serializinghtml==2.0.0",
    "sqlparse>=0.5.5",
    "stack-data==0.6.3",
    "swapper==1.4.0",
    "tomli>=2.4.0",
    "traitlets==5.14.3",
    "typing-extensions>=4.15.0",
    "urllib3==2.6.3",
    "wcwidth>=0.5.3",
    "webencodings==0.5.1",
    "whoosh==2.7.4",
]

[tool.uv.sources]
django-messages = { git = "https://github.com/frankystone/django-messages.git", rev = "8f6605be2e4f4c05817ed9922c0011aaedf18f57" }
django-tagging = { url = "https://github.com/jazzband/django-tagging/releases/download/0.5.1/django_tagging-0.5.1-py3-none-any.whl" }
